requirements.txt
psycopg2==2.9.1
python-dotenv==0.19.1
sqlparse==0.4.2
tabulate==0.8.9
yoyo-migrations==7.3.2
